Development/EcommerceMicroServices/user-service/app/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
python-multipart 
bcrypt
pydantic[email]
PyJWT
python-jose
